profiles_api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from profiles_api import views

router = DefaultRouter()
router.register('Hello-viewset', views.HelloViewSet, basename='hello-viewset')
router.register('Profile', views.UserProfileViewSet)
router.register('Feed', views.UserProfileFeedViewSet)

urlpatterns = [
    path('hello-view/', views.HelloApiView.as_view()),
    path('login/', views.UserLoginApiView.as_view()),
    path('', include(router.urls))
]
