requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.12
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
